Refs #21095: split discovery server example
